requirements.txt
fastapi[all]==0.110.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
gunicorn==22.0.0
numpy==1.26.4
opencv-python-headless==4.7.0.72
Pillow==10.3.0
tritonclient[all]==2.45.0
sentry-sdk==2.5.1
requests==2.32.3
scikit-image==0.23.2
pydantic==2.7
imageio==2.34.1
blend-modes==2.1.0
PyWavelets==1.6.0
scipy==1.13.
scikit-learn==1.5.0
catboost==1.2.5
pandas==2.2.2
redis==5.0.6
imutils==0.5.4
boto3==1.34.126
pyclipper==1.3.0.post5
shapely==2.0.4
ftfy==6.2.0
regex==2024.5.15
Wand==0.6.13
svglib==1.5.1
awscli==1.33.8
dataclasses-json==0.5.9
elastic-apm==6.22.3
diskcache==5.6.3
pyexiv2==2.12.0
protobuf==4.25.3
botocore==1.34.126
pydub==0.25.1
tqdm==4.66.4